backend/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
fastapi
uvicorn
google-genai
python-dotenv
